Create CompE 375 Lab 1
The program blinks the LED light on the ATmega328PB. A while loop is used to make the blinking infinite. The purpose of this lab was to ensure the microcontroller works and we get used to working with AtmelStudio. Data Direction Register were shifted to set the LED as an output and portB was shifted to turn the light on and off.  
The final results ended up with the ATmega328PB LED light blinks infinitely in 100ms increments. It was successful in teaching me how to use Atmel and ensured the microcontroller works.