Create CompE 375 Lab 4
Lab 4’s purpose was to scan the 4x4 keypad and transmit the pushed button. The program is to utilize the transmitting from lab 3, utilize pull up resistors and interpret a pressed button among many. The rows are outputs and columns are set to be inputs. The rows are set to low and then each column is checked to see which button is pressed. 

The result is the correct character being transmitted when it is pressed on the 4x4 keypad.