Create CompE 375 Lab 5
Lab 5’s purpose is to practice timers and counters. The check point was to have the LED slowly brighten up when the button is pressed and dim when the button is let go.  When the button is pressed the LED is going to brighten up and dim back down in 10ms. The brighten and dim comes from generating a PWM waveform, so a duty cycle occurs. 
The result is a pushed button and a LED that brightens in 5ms and dims in 5ms resulting in a total of 10ms duty cycle.