Create CompE 375 Lab 6
Lab 6’s purpose is to turn the keypad into a musical instrument. Each button that is pressed is supposed to generate a different frequency and thus note. We used the equation 1/(440 * (2^(1/12))^x ) ) to get the frequency. The checkpoint was to use basic delays. Timers were implemented after the checkpoint.
The result is that every button pushed generated a different frequency so the notes were getting to a higher pitch as each button was pressed from 1 to D.