Create CompE 375 Lab 7
Lab 7’s purpose is to introduce interrupts. The pin change interrupt was utilized to toggle the LED. The keypad was scanned to affect the duty cycle. The keypad scanning was put in the interrupt service routine. If a 1 was pressed, the LED would be at 10% brightness, 5 for 50% etc. 

The code works fine. The PWM and interrupt period was calculated according to my RedID. Every number that is pressed displays the correct brightness with A, B, C, D, #, 0, and * = to 0%